requirements.txt
graphviz==0.16
flake8==3.8.4
pytest==6.2.2
libcst==0.3.17
astpretty
apache-beam==2.27.0
coverage==5.5
ujson==4.0.2
confluent-kafka==1.6.1
beam-nuggets==0.17.1
httplib2==0.15.0
google-api-python-client==1.7.8
pytest-mock==3.5.1
PyHamcrest==2.0.2
pytest-docker-fixtures==1.3.11
pytest-timeout==1.4.2
apache-flink==1.13.0
python-dynamodb-lock-whatnick==0.9.3
pynamodb==5.0.3
boto3==1.17.93
fastapi==0.65.2
uvicorn==0.14.0
aiokafka==0.7.1